Improve alignment (#456)
* updates gcsa2

* add experimental plan to paper

* clean up alignment visualization

* output cluster alignments for mapping debug output

* remove commented-out code

* query the xg index using a set of GAM alignments

This returns the graph covered by the alignments. Useful for viz in the large.

* correctly merge patches

* detect overlapping alignments

These alternative local alignments aren't really dispersed in the sense that
the other components of the mapper assume.

For instance, global mapping quality makes no sense for them.

* use the edit length to get a distance estimate rather than XG

* set the default doubling steps to 3 (e.g. 16->128)

* begin cleanup of pair MEM restriction

* correct alignment position translation on reverse stranded alignments

* commented-out debugging experiment to understand alignment issues

* allow reading interleaved GAM with pairs

* change --mem-threading flag to -a

* resolve more bugs with MEM threader

It's now doing better than the standard clustering method!

* allow specification of paired end fragment length std deviation

Use a normal distribution.

* fix failure in compare mode (returned perfect when aln score = 0)

* expose max mem length in the functional interfaces of the mapper

This will allow re-calling methods on the same mapper with different max mem
lengths, which is a strategy I hope to use to improve the sensitivity of the
paired end mapper.

* allow sense-step to talk about MEMs or kmers, set default 5, cleanup

* un-bug the non-mem-threading bits of vg and resolve test failures

* be clear about what kind of lambda this is

Trying to make the build happy. I'm on gcc 5.4.0-6ubuntu1~16.04.2 so maybe I'm
using features that are too new.